fix: match current_offset type (uint64_t) in platform minidump writer structs
The previous commit widened current_offset to uint64_t in the base
struct but not in the Linux/macOS platform structs, causing a layout
mismatch that corrupted the crash_ctx pointer on every write_data call.

+++ src/backends/native/minidump/sentry_minidump_linux.c
@@ -86,7 +86,7 @@ typedef struct {
 typedef struct {
     // Base fields (must match minidump_writer_base_t layout)
     int fd;
-    uint32_t current_offset;
+    uint64_t current_offset;
 
     // Linux-specific fields
     const sentry_crash_context_t *crash_ctx;
